requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
chromadb
python-dotenv
pymupdf
requests